requirements.txt
python-telegram-bot>=21.0
python-dotenv
googlesearch-python
requests
aiohttp
beautifulsoup4
lxml
